Fixed the apparent bugs in the constraint solver.  It now appears to give reasonable solutions.  I also fixed a few miscellaneous bugs in the labelchecker, and our compiler caught a bug in RSAPublicKey.jif.
The login example now appears to labelcheck without problems, so the system produces labeled abstract syntax which might be splittable.  As predicted, many of the labels stay at top security, so we still have to address that problem.  There are undoubtedly still bugs in the frontend and many features still aren't supported.  I'll be working on that over the next few days.