Initial implementation of an ExtensionFactory.  All AST extension objects should be created through the factory.  jif.cup has been modified to use the factory; java12.cup and polyj.cup have not, but they should still work.  Many classes were changed to either use the factory or to pass it around so that some other class can use it.  It's almost as ubiquitous as TypeSystem--maybe we just should make both accessible through some static method and stop passing them around.
Also added an override facility.  If an extension implements one of:

 	CleanupSignaturesOverride
 	ReadSymbolsOverride
 	RemoveAmbiguitiesOverride
 	TranslateBodyOverride
 	TranslateOverride
 	TypeCheckOverride

then the corresponding function is implemented by the extension and not by
the AST node containing it.  Dispatch to the extension object is supported
in all visitor implementations in the java and jif packages.  PolyJ is
soon.  Dispatch to the extension for translate is not implemented,
but will be RSN.