Added -symbols <class> switch to jlgen, because we need to know the class name that will eventually be used with CUP (for multiple start symbols).
That symbol in the declaration of multiple start symbols is useless:
	start with <nonterminal> <func> <some symbol>;
because we can just generate a new <symbol> and be guaranteed it's unique.
Besides, why force the user to come up with new symbols? We can do it
ourselves. So now we do. Syntax is cleaner:
	start with <nonterminal> <func>;

Documentation for the two items above is coming right up.

Fixed up some Makefiles in tests and JLgen.