Merged QualfiedNew with New.  Fixed type checking and disambiguation for New.
Fixed bug in ClassContextResolver which allowed a class to be
qualified by itself, e.g. C.C.C is the same as C.

Changed interface for Type.translate and Label.translate to take
a Resolver rather than a Context.  This allows qualified new to
be translated correctly.  This signature change accounts for most
of the files below.