First cut at allowing multiple ASTs for multiple languages to live in compiler at the same time.  The major changes were:
Visitors don't get ask the compiler for their type system and
node factory; instead, these are passed in to the visitor by the
ExtensionInfo.

There is no longer one ImportTable per Job.  The ImportTable is
now associated with a SourceFile node.

Added a SourceCollection node so that multiple source files can
live in the same AST.

Various changes to visitor constructors to clean up some
ugliness.  Removed BaseVisitor--it had some extra functionality
that is unnecessary for many visitors, but was being used anyway
because, due to its name, it was thought to be the base class of
all visitors.

The JifTypeSystem now uses a vanilla-Java TypeSystem to load
external principals since they are implemented as Java class
files, not Jif class files.  This permitted me to remove an ugly
hack in JifContext.

This change is still not how I want it: we can still only have
one _source_ language since all Jobs must be created by the same
ExtensionInfo so barrier passes work correctly.  I don't think
this is a big limitation right now.

Next step is to implement a rewriting pass so we can start
rewriting polymix (or whatever the mixin language will be
called) ASTs to polyj ASTs.