Support for multiple extension nodes, that is, extensions can now have extensions.  Also got rid of the requirement that methods be called through the extension, e.g., n.ext().typeCheck(); the Node_c class handles dispatching to the extension so you can now just call n.typeCheck(). This change was made so that you don't have to chase ext() pointers to get to the outermost extension.
Another change: the Node interface now extends the Ext
interface.  This allows a Node to be its own Ext, which is
now what happens by default.  This should save a little
memory and maybe some time.

All the interesting changes are in Ext_c and Node_c.
But, most of the changes below are to get rid of the useless
ext() calls.