Implemented end line and columns for Positions, as per Ondrej's suggestion.
The lexer is responsible for assigning appropriate end lines and columns to tokens. If the lexer does not do this, then the parser behaves as it did previously, assigning positions to terms the same way it did previously; if the lexer does assign end lines and columns, then the parser sets the positions of terms appropriately for each kind of term (done in java12.cup).

There is an issue with how to display errors where the position goes over multiple lines. Also, some of the choices for what exactly constitutes the beginning and end of terms may need to be modified slightly.