Serialization bug fixes:
- Member classes should not be serialized in
the same class file as their container.  Doing so may cause
null pointer exceptions via a complicated series of events
during deserialization.

- Save and restore the system resolver during deserialization to
prevent partially reconstructed types from leaking out.

- Deserialized classes need initialization before being returned
from the system resolver to add themselves to their container
(and in Jx to setup some scheduler dependencies).  Added LazyInitializer
interface that is called to initialize the types when types are added
to the system resolver.

- Disabled MemberClassResovler.  Instead, ClassContextResolver
will look for the class name A$B as well as A.B when looking for
a member class.

Jx translation bugs:
- Don't rename interface methods; there may not
be a single interface that introduces the method.  This changes the
semantics of JxMI: interfaces now never conflict.

- Use the same mechanism for both class and package member lookup at run time.
Necessary since member classes are no longer serialized with their container.
Also simplifies the code.

Cleaned up instantiation code in JxTypeSytsem, removing much duplication.

Refactored some of the member class/package lookup code in Jx.