Added an option -outputamb to allow output directly from ambiguous nodes. Ambiguous nodes now check for the flag, and only throw exceptions when the flag is not set.
Together with -c, it is very fast now. Only 3 minutes for compiling composed lambda compiler on my mac, compared to somewhere between 1 and 2 hours on my thinkpad, which is a faster computer, before this change.

Note: the jl part of the compiler is much faster (at least 2x) with java 1.5 on windows machines. However, covariant return values in 1.5 library cause problems for compiling CodeWriter. So I have to set back with 1.4.2, which seems to have slower iterators for LinkedList and TreeMap.